==> epigenomic_dna_methylation/GSE193535_preprocessing/1-load_data.R <==
########### 1 - LOAD DATA 

# set up the path to directory with raw methylation files (IDAT)
dataDirectory <- "../input"
# check the files are there
list.files(dataDirectory, recursive=TRUE)

# load necessary libraries
library(minfi)
library(IlluminaHumanMethylation450kanno.ilmn12.hg19)
library(IlluminaHumanMethylation450kmanifest)

# get 450k annotation data
ann450k = getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19)
# check the data
head(ann450k)

# read in the sample sheet
targets <- read.metharray.sheet(dataDirectory, pattern="methylation2_sample_sheet.csv")

# read in the raw data from the IDAT files
rgSet <- read.metharray.exp(targets=targets, force=TRUE)
rgSet

# give the samples descriptive names e.g. sample type and sample name
targets$ID <- paste(targets$Sample_Type, targets$Sample_Name, sep = ".")
sampleNames(rgSet) <- targets$ID
rgSet

# save outputs
saveRDS(rgSet, file = "intermediate/rgSet.rds")
saveRDS(targets, file = "intermediate/targets.rds")
saveRDS(ann450k, file = "intermediate/ann450k.rds")
